Big keccakf cleanup
