+++ precompiles/helpers/src/keccak/keccak_f/mod.rs
@@ -3,7 +3,7 @@ mod round;
 mod utils;
 
 use constants::*;
-use round::keccak_f_round;
+pub use round::keccak_f_round;
 pub use utils::*;
 
 /// State representation as 5x5x64 bits
@@ -24,77 +24,27 @@ fn reduce_state_mod2(state: &mut KeccakStateBits) {
     state.iter_mut().flatten().flatten().for_each(|bit| *bit %= 2);
 }
 
-/// Iterator that yields the state after each round of Keccak-f
-pub struct KeccakRoundIterator {
-    state: KeccakStateBits,
-    round: usize,
-    initial_returned: bool,
-}
-
-impl KeccakRoundIterator {
-    pub fn new(initial_state: KeccakStateBits) -> Self {
-        Self { state: initial_state, round: 0, initial_returned: false }
-    }
-}
-
-impl Iterator for KeccakRoundIterator {
-    type Item = (KeccakStateBits, usize); // (state, round_number)
-
-    fn next(&mut self) -> Option<Self::Item> {
-        // First return the initial state as round 0
-        if !self.initial_returned {
-            self.initial_returned = true;
-            return Some((self.state, 0));
-        }
-
-        // Check if we've completed all 24 rounds
-        if self.round >= 24 {
-            return None;
-        }
-
-        // Perform one round of Keccak-f
-        keccak_f_round(&mut self.state, self.round);
-
-        // Return the unreduced state (before modulo 2)
-        let unreduced_state = self.state;
-
-        // Apply reduction for the next round
-        reduce_state_mod2(&mut self.state);
-
-        self.round += 1;
-
-        Some((unreduced_state, self.round))
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        let remaining = if self.initial_returned {
-            24 - self.round
-        } else {
-            25 - self.round // Initial state + 24 rounds
-        };
-        (remaining, Some(remaining))
-    }
-}
-
-/// Function-based iterator for Keccak-f rounds  
-pub fn keccak_f_rounds(initial_state: KeccakStateBits) -> KeccakRoundIterator {
-    KeccakRoundIterator::new(initial_state)
-}
-
-/// Iterator that yields just the states (without round numbers)
-pub fn keccak_f_round_states(
-    initial_state: KeccakStateBits,
-) -> impl Iterator<Item = KeccakStateBits> {
-    keccak_f_rounds(initial_state).map(|(state, _)| state)
-}
-
 #[cfg(test)]
 mod tests {
-    use super::{
-        keccak_f, keccak_f_round_states, reduce_state_mod2,
-        utils::{keccakf_state_from_linear, keccakf_state_to_linear},
-        KeccakRoundIterator,
-    };
+    use super::{keccak_f, utils::keccakf_state_from_linear, KeccakStateBits};
+
+    /// Convert from 5x5x64 bit array to linear [u64; 25]
+    #[allow(clippy::needless_range_loop)]
+    pub fn keccakf_state_to_linear(state: &KeccakStateBits) -> [u64; 25] {
+        let mut linear = [0u64; 25];
+        for x in 0..5 {
+            for y in 0..5 {
+                let mut word = 0u64;
+                for z in 0..64 {
+                    if state[x][y][z] == 1 {
+                        word |= 1u64 << z;
+                    }
+                }
+                linear[x + y * 5] = word;
+            }
+        }
+        linear
+    }
 
     #[test]
     fn test_keccak_f_zero_state() {
@@ -237,61 +187,5 @@ mod tests {
                 0x20D06CD26A8FBF5C,
             ]
         );
-    }
-
-    #[test]
-    fn test_keccak_f_round_iterator() {
-        let initial_state_linear = [0u64; 25];
-        let initial_state = keccakf_state_from_linear(&initial_state_linear);
-        let round_iter = KeccakRoundIterator::new(initial_state);
-
-        // Test that we get exactly 25 rounds
-        let all_rounds: Vec<_> = round_iter.collect();
-        assert_eq!(all_rounds.len(), 25);
-
-        // Test that round numbers are correct
-        for (i, (_, round_num)) in all_rounds.iter().enumerate() {
-            assert_eq!(*round_num, i);
-        }
-
-        // Test that round 0 is the initial state
-        let (round_0_state, round_0_num) = &all_rounds[0];
-        assert_eq!(*round_0_num, 0);
-        let round_0_linear = keccakf_state_to_linear(round_0_state);
-        assert_eq!(round_0_linear, initial_state_linear);
-
-        // Test that the final state (after reduction) matches the full keccak_f result
-        let final_unreduced_state = all_rounds.last().unwrap().0;
-
-        // Apply reduction to the final unreduced state to compare with keccak_f result
-        let mut final_reduced_state = final_unreduced_state;
-        reduce_state_mod2(&mut final_reduced_state);
-        let final_reduced_linear = keccakf_state_to_linear(&final_reduced_state);
-
-        let mut expected_state = keccakf_state_from_linear(&initial_state_linear);
-        keccak_f(&mut expected_state);
-        let expected_linear = keccakf_state_to_linear(&expected_state);
-
-        assert_eq!(final_reduced_linear, expected_linear);
-    }
-
-    #[test]
-    fn test_keccak_f_round_states() {
-        let initial_state_linear = [0u64; 25];
-        let initial_state = keccakf_state_from_linear(&initial_state_linear);
-        let states: Vec<_> = keccak_f_round_states(initial_state).collect();
-
-        assert_eq!(states.len(), 25);
-
-        // Final state should match full keccak_f (after reduction)
-        let mut final_unreduced_state = states[24];
-        reduce_state_mod2(&mut final_unreduced_state);
-        let final_reduced_linear = keccakf_state_to_linear(&final_unreduced_state);
-
-        let mut expected_final = keccakf_state_from_linear(&initial_state_linear);
-        keccak_f(&mut expected_final);
-        let expected_linear = keccakf_state_to_linear(&expected_final);
-
-        assert_eq!(final_reduced_linear, expected_linear);
     }
 }
