+++ precompiles/helpers/src/keccak/keccak_f/utils.rs
@@ -25,24 +25,6 @@ pub fn keccakf_state_from_linear(linear: &[u64; 25]) -> KeccakStateBits {
     state
 }
 
-/// Convert from 5x5x64 bit array to linear [u64; 25]
-#[allow(clippy::needless_range_loop)]
-pub fn keccakf_state_to_linear(state: &KeccakStateBits) -> [u64; 25] {
-    let mut linear = [0u64; 25];
-    for x in 0..5 {
-        for y in 0..5 {
-            let mut word = 0u64;
-            for z in 0..64 {
-                if state[x][y][z] == 1 {
-                    word |= 1u64 << z;
-                }
-            }
-            linear[x + y * 5] = word;
-        }
-    }
-    linear
-}
-
 #[allow(clippy::needless_range_loop)]
 pub fn keccakf_state_to_linear_1d(state: &KeccakStateBits) -> [u64; 1600] {
     let mut linear_1d = [0u64; 1600];
@@ -55,15 +37,6 @@ pub fn keccakf_state_to_linear_1d(state: &KeccakStateBits) -> [u64; 1600] {
         }
     }
     linear_1d
-}
-
-pub const fn keccakf_idx_pos(idx: usize) -> (usize, usize, usize) {
-    debug_assert!(idx < 1600);
-
-    let x = (idx / 64) % 5;
-    let y = (idx / 320) % 5;
-    let z = idx % 64;
-    (x, y, z)
 }
 
 pub const fn keccakf_bit_pos(x: usize, y: usize, z: usize) -> usize {
