+++ precompiles/helpers/src/keccak/mod.rs
@@ -1,7 +1,5 @@
 mod keccak_f;
 mod keccak_f_expr;
-mod keccak_f_state;
 
 pub use keccak_f::*;
 pub use keccak_f_expr::keccak_f_expr;
-pub use keccak_f_state::keccak_f_state;
